% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMR.R
\name{BMR}
\alias{BMR}
\title{Find background mutation rate}
\usage{
BMR(preOutM, preOutC, preOutV, bmr = 1.2e-06)
}
\arguments{
\item{preOutM}{Mutation data output by preprocessing function or designated by users.}

\item{preOutC}{Coverage data output by preprocessing function or designated by users.}

\item{preOutV}{Covariate data which includes expression level, replication time and chromatin compartment defaultly, and could be altered by users.}

\item{bmr}{The default background mutation rate is 1.2e-6, and the value alters when function ends.}
}
\value{
The output BMR_out is an input to the sigGenes function.

The output BMR_out is an input to the sigGenes function.
}
\description{
It uses the data output by the preprocessing function in order to get
background mutation rate. It uses the method of bagel. The result will be then
used in varified specific gene detecting methods.
}
\author{
Xiaolu Xu <lu.xu@lnnu.edu.cn>

It uses the data output by the preprocessing function in order to get
background mutation rate. It uses the method of bagel. The result will be then
used in varified specific gene detecting methods.

Xiaolu Xu <lu.xu@lnnu.edu.cn>
}
