% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocument.R
\docType{data}
\name{BMR_out}
\alias{BMR_out}
\title{Output of BMR function}
\format{
A list of 11 elements:
\describe{
 \item{G}{Specific gene}
 \item{N_silent}{The number of silent coverages}
 \item{n_silent}{The number of silent mutations}
 \item{N_nonsilent}{The number of nonsilent coverages}
 \item{n_nonsilent}{The number of nonsilent mutations}
 \item{N_noncoding}{The number of noncoding coverages}
 \item{n_noncoding}{The number of noncoding mutations}
 \item{X_gcp}{Used in gene detecting methods}
 \item{x_gcp}{Used in gene detecting methods}
 \item{null_categ}{The number of categories}
 \item{nanalysis}{Number of rows}
 }
}
\usage{
data(SampleBMRout)
}
\description{
A list containing result of BMR method
}
\keyword{datasets}
