% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocument.R
\docType{data}
\name{C}
\alias{C}
\title{Sample coverage data}
\format{
A data frame with 31104 rows and 4 variables:
\describe{
 \item{gene}{The gene of mutation}
 \item{effect}{The effect of mutation}
 \item{categ}{The specific category}
 \item{coverage}{The number of coverage}
 }
}
\usage{
data(SampleCoverage)
}
\description{
A data set containing raw information of coverages
}
\keyword{datasets}
