% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocument.R
\docType{data}
\name{dict}
\alias{dict}
\title{Mutation type dictionary}
\format{
A data frame with 50 rows and 2 variables:
\describe{
 \item{Variant_Classification}{Variant types of effect}
 \item{effect}{Specified types for use}
 }
}
\usage{
data(SampleDict)
}
\description{
A data set containing variant mutation types and their correspondings for use
}
\keyword{datasets}
