% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DriverPathway.R
\name{DriverPathway}
\alias{DriverPathway}
\title{Identify driver pathway using a de novo method, AWRMP(https://www.frontiersin.org/articles/10.3389/fgene.2019.00233/full).}
\usage{
DriverPathway(
  mutation_data,
  driver_size = 3,
  pop_size = 200,
  iters = 1000,
  permut_time = 1000,
  process_bmr = 1.2e-06
)
}
\arguments{
\item{mutation_data}{0-1 mutation matrix where rows represent patients, columns represent genes or MAF file.
If the mutation data is a MAF file, then the preprocessing procedure will be performed using coverage data, covariate data, mutation dictionary, and chr files directory, which can be specified by the user.
The user should name these files "coverage.txt.$" for coverage data, "covariates.txt$" for covariate data, "dictionary file.txt$" for mutation dictionary, and "^chr files hg" for "chr files directory".
If the specified named files can be found, they will be read and utilized to generate mutation matrix.
If the files cannot be detected, the function will automatically download them from the MutSigCV website, which may take some time.}

\item{driver_size}{The size of identified driver gene set, defaulted to 3.}

\item{pop_size}{The population size of GA which should be adjusted according the number of samples, defaulted to 200.}

\item{iters}{The iteration time of GA, defaulted to 1000.}

\item{permut_time}{The times of permutation test, defaulted to 1000.}

\item{process_bmr}{The background mutation rate to preprocess MAF file to mutation matrix using binomial hypothesis testing.}
}
\value{
A list including the mutation matrix, identified driver gene set, and p-value.
}
\description{
Identify driver pathway using a de novo method, AWRMP(https://www.frontiersin.org/articles/10.3389/fgene.2019.00233/full).
}
\examples{
data(SampleMutationMatrix)
driverSet <- DriverPathway(mutation_matrix, driver_size=3, pop_size=20, iters=50, permut_time=100)
}
\author{
Xiaolu Xu <lu.xu@lnnu.edu.cn>
}
