% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocument.R
\docType{data}
\name{M}
\alias{M}
\title{Sample mutation data}
\format{
A data frame with 816 rows and 12 variables:
\describe{
 \item{Hugo_Symbol}{The gene where the mutation happens}
 \item{Chromosome}{The Chromosome where the mutation happens}
 \item{Start_position}{The start position of the mutation}
 \item{End_position}{The end position of the mutation}
 \item{Variant_Classification}{Variant type of mutation}
 \item{Reference_Allele}{Reference allele}
 \item{Tumor_Seq_Allele1}{The first tumor allele}
 \item{Tumor_Seq_Allele2}{The second tumor allele}
 \item{Tumor_Sample_Barcode}{Code name of the patient}
 \item{is_coding}{If the mutation is in the coding area}
 \item{is_silent}{If the mutation is in the silent area}
 \item{categ}{The initial category of the mutation}
 }
}
\usage{
data(SampleMutationMaf)
}
\description{
A Maf data set containing raw information of mutations
}
\keyword{datasets}
