% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocument.R
\docType{data}
\name{mutation_matrix}
\alias{mutation_matrix}
\title{Mutation matrix}
\format{
An object of class \code{data.frame} with 50 rows and 50 columns.
}
\usage{
data(SampleMutationMatrix)
}
\description{
A matrix of mutations occur in which patients and genes.
The values are either 1 or 0, representing mutated or not.
}
\keyword{datasets}
