% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocument.R
\docType{data}
\name{preOutC}
\alias{preOutC}
\title{Preprocessed coverage data}
\format{
A data frame with 810 rows and 4 variables:
\describe{
 \item{gene}{The gene of mutation}
 \item{effect}{The effect of mutation}
 \item{categ}{The specific category}
 \item{coverage}{The number of coverage}
 }
}
\usage{
data(SampleOutCoverage)
}
\description{
The output coverage data of preprocessing
}
\keyword{datasets}
