% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preSmallFunctions.R
\name{preprocessAssignCateg}
\alias{preprocessAssignCateg}
\title{Assigning categories for mutation types}
\usage{
preprocessAssignCateg(K)
}
\arguments{
\item{K}{Dataframe of previously detected mutation categories}
}
\description{
For each of the 192 mutation types, assigning a corresponding category from number 1 to 4.
}
\author{
Xiaolu Xu <lu.xu@lnnu.edu.cn>
}
