% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preSmallFunctions.R
\name{preprocessCollapseNn65to32}
\alias{preprocessCollapseNn65to32}
\title{Collapsing category number to 32}
\usage{
preprocessCollapseNn65to32(Y)
}
\arguments{
\item{Y}{Dataframe of numbers of coverage and 4 mutation types of 65 categories}
}
\description{
Collapsing previous category number from 65 to 32.
}
\details{
This function collapses the previous category number from 65 to 32 by first removing the "any N" type,
then fusing a specific trinucleotide and its opposite.
}
