% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preSmallFunctions.R
\name{preprocessConcatStructsKeepFields}
\alias{preprocessConcatStructsKeepFields}
\title{Appending a new catagory to the previous ones}
\usage{
preprocessConcatStructsKeepFields(K, K2)
}
\arguments{
\item{K}{Dataframe of previously detected mutation categories}

\item{K2}{Dataframe of the newly defined category}
}
\description{
Adding a new category representing 'null+indel' to the previously detected 4 categories.
}
\details{
The number of categories is set 4 defaultly, this function append the newly defined 'null+indel' category
to the previous ones.
}
