% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preSmallFunctions.R
\name{preprocessGenerateCategContext65}
\alias{preprocessGenerateCategContext65}
\title{Generating 65 mutation categories}
\usage{
preprocessGenerateCategContext65()
}
\description{
Generating 65 types of mutaion categories for subsequently classification and analysis.
}
\details{
Considering the contexts for each of the 4 bases(A,C,G,T), i.e. the total number of categories reaches
to 65(64 types of trinucleotides and a supplement called "any N").
}
