% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocessing}
\alias{preprocessing}
\title{Preprocess input data}
\usage{
preprocessing(
  M,
  C,
  dict,
  V,
  chr_files_directory,
  categ_flag = NaN,
  output_filestem = "Preprocessed"
)
}
\arguments{
\item{M}{Mutation maf data.}

\item{C}{Coverage raw data.}

\item{dict}{Mapping Variant_Classification to mutation effect in mutation data.}

\item{V}{Covariate raw data.}

\item{chr_files_directory}{Chromosome files directory, hg19 or hg18.}

\item{categ_flag}{Mutation category number, should be either NaN or numeric.}

\item{output_filestem}{Unified prefix of output data, could be modified as well.}
}
\value{
The output is a list includes the preprocessed mutation and coverage data, which are the inputs of BMR function.
}
\description{
This function depends on a chromosome files folder, hg19 or hg38, which includes the txt files of each chromosome.
The hg19 folder is able to download at the url of our Github https://github.com/bioinformatics-xu/DriverGenePathway.
This function preprocesses the mutation data and coverage data and selects important mutation categories.
It generates intermediate results including preprocessed mutation file, coverage file, categ file and mutcateg_discovery file.
}
