% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigGenes.R
\name{sigGenes}
\alias{sigGenes}
\title{Identify significante genes}
\usage{
sigGenes(
  BMR_out,
  p_class = "allTest",
  output_filestem = "output",
  sigThreshold = 0.05
)
}
\arguments{
\item{BMR_out}{Data output by the BMR function.}

\item{p_class}{"BB" represents beta binomial distribution test; "FCPT" represents Fisher combined P-value test;
"LRT" represents likelihood ratio test; "CT" represents convolution test; "PJ" represents projection test
method; "allTest" represents the mutual results of all methods.}

\item{output_filestem}{Unified prefix of output data, could be modified as well}

\item{sigThreshold}{The threshhold of q-value to judge if the gene is significant}
}
\value{
Signifiant genes
}
\description{
Applying different hypothesis test methods to identify significant genes on output
of the BMR function
}
\details{
This function searches the significant genes using different hypothesis test methods,
including binomial distribution test, beta binomial distribution test, Fisher combined P-value test,
likelihood ratio test and convolution test.
}
