% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocument.R
\docType{data}
\name{V}
\alias{V}
\title{Covariate data}
\format{
A data frame with 54 rows and 4 variables:
\describe{
 \item{gene}{Specific gene}
 \item{expr}{Expression level}
 \item{reptime}{Replication time}
 \item{hic}{Chromatin compartment}
 }
}
\usage{
data(SampleCovariate)
}
\description{
The covariate data used for background mutation rate discovery
}
\keyword{datasets}
