% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DriverGene.R
\name{DriverGene}
\alias{DriverGene}
\title{Identify driver genes}
\usage{
DriverGene(
  Mutation = NULL,
  Coverage = NULL,
  Covariate = NULL,
  MutationDict = NULL,
  chr_files_directory = NULL,
  categ_flag = NaN,
  bmr = 1.2e-06,
  p_class = "BB",
  sigThreshold = 0.05,
  output_filestem = "output"
)
}
\arguments{
\item{Mutation}{Mutation maf data, mandatory data.}

\item{Coverage}{Coverage raw data, read from exome_full192.coverage.txt in MutSigCV by default.}

\item{Covariate}{Covariate data, read from "gene.covariates.txt" in MutSigCV by default.}

\item{MutationDict}{Mutation dictionary to map Variant_Classification to mutation effect in mutation data, read from "mutation_type_dictionary_file.txt" in MutSigCV by default.}

\item{chr_files_directory}{Chromosome files directory, hg19 or hg38.}

\item{categ_flag}{Mutation category number, should be either NaN or numeric, defaulted to NaN.}

\item{bmr}{The default background mutation rate is 1.2e-6, and the value alters when
function ends.}

\item{p_class}{Hypothesis test methods. "BB" represents beta binomial distribution test; "FCPT" represents Fisher combined P-value test;
"LRT" represents likelihood ratio test; "CT" represents convolution test; "projection" represents projection test
method; "allTest" represents the mutual results of all methods.}

\item{sigThreshold}{The threshhold of q-value to judge if the gene is significant.}

\item{output_filestem}{The parameters to name the output files, defaulted to "output".}
}
\description{
This function depends on a chromosome files folder, hg19 or hg38, which includes the txt files of each chromosome.
The hg19 folder is able to download at the url of our Github https://github.com/bioinformatics-xu/DriverGenePathway.
This function first preprocesses input data via the preprocessing function. Then the BMR function is used to calculate
the background mutation rate. In order to identify driver genes, there are five hypothesis test methods to select
from according to the parameter p_class, i.e. binomial distribution test, represents beta binomial distribution test,
Fisher combined P-value test, likelihood ratio test, convolution test and 2D-projection method, which are
all encapsulated in the sigGenes function. The output files contains txt files including the preprocessed
mutation and coverage data, mutation categories, and significant driver genes. There are several plots output as
pdf files as well.
}
\author{
Xiaolu Xu <lu.xu@lnnu.edu.cn>
}
