% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DriverPathway.R
\name{DriverPathway}
\alias{DriverPathway}
\title{Search driver pathway using de novo method}
\usage{
DriverPathway(
  mutation_data,
  driver_size = 3,
  pop_size = 200,
  iters = 500,
  permut_time = 1000,
  process_bmr = 1.2e-06,
  outfile = "denovoDriverPathway.txt"
)
}
\arguments{
\item{driver_size}{The size of identified driver gene set, defaulted to 3.}

\item{pop_size}{The population size of GA, defaulted to 200.}

\item{iters}{The iteration time of GA, defaulted to 500.}

\item{permut_time}{The times of permutation test, defaulted to 1000.}

\item{process_bmr}{The background mutation rate for preprocessing mutation matrix using MutsigCV1.0.}

\item{outfile}{The output file of driver gene set.}

\item{mutation_matrix}{0-1 mutation matrix where rows represent patients, columns represent genes.}
}
\description{
This function searches driver pathway using de novo method based on mutual exclusivity and coverage
It outputs the driver pathway and its p-value as a txt file.
}
\examples{
data(SampleMutationMatrix)
DriverPathway(mutation_matrix, driver_size=3, pop_size=20, iters=50, permut_time=100,process_bmr=1.2e-6)
}
\author{
Xiaolu Xu <lu.xu@lnnu.edu.cn>
}
